TST: port some SeqView tests to new_sequence.py
